srtp: implement rekeying
This change introduces the ability to switch SRTP (not SRTCP) keys while a stream is running.
It adds 3 methods to media_stream to update sending and receiving keys (exchange out-of-band).

On the sender side, the SRTP context is simply recreated with the new keys immediately.
For the receiver, we need to account for the packets encrypted and authenticated with the old
keys that are still in transit even after the receiver has been told to use new keys.
To solve this, rekeying the receiving side actually keeps 2 active SRTP contexts, allowing
packets protected with either of the two key sets to be decrypted.
After a while (decided by the user), the old SRTP context is discarded and only the new key
is kept.
Note that rekeying only works if HMAC-SHA1 authentication tags are used, otherwise we can't
know which key to use for packets.